requirements.txt
torch
transformers
bitsandbytes
peft
accelerate
datasets
python-dotenv
protobuf
sentencepiece
scikit-learn